IRGen: Split hoisting tests out of weak_import_native.swift into a new test case.
Once the `CHECK-DAG` directives in `weak_import_native` were being recognized correctly, the `CHECK:` lines later in the test were interfering with `CHECK-DAG` matches because some of them straddled the existing `CHECK:` lines.
